# api/api.yml
openapi: '3.0.3'
info:
  title: Piperchat API
  version: '1.0'
  description: |
    Piperchat design API description
...
paths:
...
  /users/{username}/photo:
    get:
      summary: Get photo
      description: Get the photo of a user
      tags:
        - 'users'
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/username'
      responses:
        '200':
          '$ref': '#/components/responses/UserPhoto'
        '400':
          '$ref': '#/components/responses/BadRequest'
        '401':
          '$ref': '#/components/responses/JwtTokenMissingOrInvalid'
        '404':
          '$ref': '#/components/responses/UserNotFound'
  /users/{username}/description:
    get:
      summary: Get description
      description: Get the description of a user
      tags:
        - 'users'
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/username'
      responses:
        '200':
          '$ref': '#/components/responses/UserDescription'
        '400':
          '$ref': '#/components/responses/BadRequest'
        '401':
          '$ref': '#/components/responses/JwtTokenMissingOrInvalid'
        '404':
          '$ref': '#/components/responses/UserNotFound'
...
  /servers/{serverId}/channels:
    get:
      summary: Get channels list
      description: Get the list of message and multimedia channels in a server
      tags:
        - 'channels'
      security:
        - bearerAuth: []
      parameters:
        - '$ref': '#/components/parameters/serverId'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                type: array
                items:
                  '$ref': '#/components/schemas/basicChannel'
                example:
                  - channelId: channel123
                    name: Beautiful Channel
                    description: This channel is created for make new friends!
                    channelType: message
                  - channelId: channel456
                    name: Beautiful vocal channel
                    description: This channel is created for make new friends!
                    channelType: multimedia
        '401':
          '$ref': '#/components/responses/UnauthorizedError'
        '403':
          '$ref': '#/components/responses/ForbiddenError'
        '404':
          description: Server not found
    post:
      summary: Create a new channel
      description: The server owner create a new channel
      tags:
        - 'channels'
      security:
        - bearerAuth: []
      parameters:
        - '$ref': '#/components/parameters/serverId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                channelType:
                  enum:
                    - message
                    - multimedia
                name:
                  type: string
                  example: channel1
                description:
                  type: string
                  example: This channel is created for make new friends!
              required:
                - channelType
                - name
      responses:
        '200':
          description: Ok
        '400':
          description: Bad request
        '401':
          '$ref': '#/components/responses/UnauthorizedError'
        '403':
          '$ref': '#/components/responses/ForbiddenError'
        '404':
          description: Server not found
  /servers/{serverId}/channels/{channelId}:
    put:
      summary: Modify a channel
      description: The owner modify channel information
      tags:
        - 'channels'
      security:
        - bearerAuth: []
      parameters:
        - '$ref': '#/components/parameters/serverId'
        - '$ref': '#/components/parameters/channelId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Beautiful Channel 2.0
                description:
                  type: string
                  example: This channel is created for make new friends! 2.0
      responses:
        '200':
          description: Ok
        '400':
          description: Bad request
        '401':
          '$ref': '#/components/responses/UnauthorizedError'
        '403':
          '$ref': '#/components/responses/ForbiddenError'
        '404':
          description: Server or Channel not found
    delete:
      summary: Delete a channel
      description: The owner delete a channel
      tags:
        - 'channels'
      security:
        - bearerAuth: []
      parameters:
        - '$ref': '#/components/parameters/serverId'
        - '$ref': '#/components/parameters/channelId'
      responses:
        '200':
          description: Ok
        '401':
          '$ref': '#/components/responses/UnauthorizedError'
        '403':
          '$ref': '#/components/responses/ForbiddenError'
        '404':
          description: Server or Channel not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              params:
                description: List of parameters that are not validß
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: username
                    expected:
                      type: string
                      example: string
                    received:
                      type: string
                      example: number
              body:
                description: List of body fields that are not valid
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: username
                    expected:
                      type: string
                      example: string
                    received:
                      type: string
                      example: number
              query:
                description: List of query fields that are not valid
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                      example: username
                    expected:
                      type: string
                      example: string
                    received:
                      type: string
                      example: number
    JwtTokenMissingOrInvalid:
      description: JWT token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: JWT token Missing or Invalid
    UnauthorizedError:
      description: Access token is missing or invalid
    ForbiddenError:
      description: You don't have permission to access this resource
    UserNotFound:
      description: User not found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: User not found

    # LOGIN
...
    UserPhoto:
      description: Ok
      content:
        application/json:
          schema:
            type: object
            properties:
              photo:
                type: string
                format: base64
    # GET USER DESCRIPTION
    UserDescription:
      description: Ok
      content:
        application/json:
          schema:
            type: object
            properties:
              description:
                type: string
                example: 'Here to make new friends'
    # DIRECT MESSAGES
...
  schemas:
...
    basicChannel:
      type: object
      properties:
        channelId:
          type: string
          example: channel123
        name:
          type: string
          example: Beautiful Channel
        description:
          type: string
          example: This channel is created for make new friends!
...
  parameters:
    username:
      name: username
      in: path
      required: true
      description: username of the user that you want to interact with
      schema:
        type: string
      example: user1
...
    serverId:
      name: serverId
      in: path
      required: true
      description: serverId of the server that you want to interact with
      schema:
        type: string
      example: server123
    channelId:
      name: channelId
      in: path
      required: true
      description: channelID of the message channel that you want to interact with
      schema:
        type: string
      example: channel123